Rename clearFileSearch action and move to src/actions/ui.js
